Make sure AgentServer actually executes shutDown
Also install signal handlers in shutDown so if they
block, and you hit control-c again, we make sure it dies.

Normally, if Runtime.exit() gets called it runs teardown
handlers and the like, but if you call .exit again while
this happens, you get bupkis.  That's why we install the
signal handlers during the teardown, so if it blocks after
we get called, we can go home before teardown is complete.